requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
httpx==0.24.1
sqlalchemy==2.0.20
pydantic==1.10.15
apscheduler==3.9.0
lxml==4.9.3
python-dotenv==1.0.0
loguru==0.7.0
psycopg2-binary==2.9.9
requests==2.31.0
deep-translator==1.11.4
pytz==2023.3
aiogram==2.25.1